pyspark_data_transformation/requirements.txt
cassandra-driver==3.29.2
pyspark==3.5.3
python-dotenv==1.0.1